[setup] Move version switcher to its own component
